refactor(admin): extract admin domain package from auth
