+++ apps/backend/src/hrm_backend/admin/__init__.py
@@ -1,0 +1,1 @@
+"""Admin domain package for staff governance and admin-only operations."""
