+++ apps/backend/src/hrm_backend/admin/dependencies/__init__.py
@@ -1,0 +1,5 @@
+"""Dependency providers for admin domain services."""
+
+from hrm_backend.admin.dependencies.admin import get_admin_service
+
+__all__ = ["get_admin_service"]
