+++ apps/backend/src/hrm_backend/admin/infra/__init__.py
@@ -1,0 +1,1 @@
+"""Infrastructure adapters for admin domain."""
