+++ apps/backend/src/hrm_backend/admin/models/__init__.py
@@ -1,0 +1,5 @@
+"""Admin domain model namespace.
+
+Admin domain currently reuses auth persistence models for staff accounts
+and employee registration keys.
+"""
