+++ apps/backend/src/hrm_backend/admin/routers/__init__.py
@@ -1,0 +1,5 @@
+"""HTTP routers for admin domain APIs."""
+
+from hrm_backend.admin.routers.v1 import router as admin_router
+
+__all__ = ["admin_router"]
