+++ apps/backend/src/hrm_backend/admin/schemas/requests.py
@@ -1,21 +1,24 @@
-"""Request payload schemas for authentication endpoints."""
+"""Request payload schemas for admin endpoints."""
 
 from __future__ import annotations
 
-from uuid import UUID
+from typing import Literal
 
 from pydantic import BaseModel, ConfigDict, Field, field_validator
 
+StaffRoleClaim = Literal["admin", "hr", "manager", "employee", "leader", "accountant"]
 
-class RegisterRequest(BaseModel):
-    """Input payload for staff self-registration with one-time employee key."""
+
+class AdminCreateStaffRequest(BaseModel):
+    """Input payload for admin-managed direct staff account creation."""
 
     model_config = ConfigDict(extra="forbid")
 
     login: str = Field(min_length=3, max_length=64)
     email: str = Field(min_length=3, max_length=256)
     password: str = Field(min_length=12, max_length=256)
-    employee_key: UUID
+    role: StaffRoleClaim
+    is_active: bool = True
 
     @field_validator("login")
     @classmethod
@@ -33,26 +36,19 @@ class RegisterRequest(BaseModel):
         return normalized
 
 
-class LoginRequest(BaseModel):
-    """Input payload for staff login.
-
-    Supports identifier/password staff password flow.
-    """
+class AdminStaffUpdateRequest(BaseModel):
+    """Input payload for admin-managed staff role/active-state updates."""
 
     model_config = ConfigDict(extra="forbid")
 
-    identifier: str = Field(min_length=3, max_length=256)
-    password: str = Field(min_length=12, max_length=256)
-
-    @field_validator("identifier")
-    @classmethod
-    def normalize_identifier(cls, value: str) -> str:
-        """Normalize login/email identifier for account lookup."""
-        return value.strip().lower()
+    role: StaffRoleClaim | None = None
+    is_active: bool | None = None
 
 
-class RefreshRequest(BaseModel):
-    """Input payload for access/refresh token rotation."""
+class AdminCreateEmployeeKeyRequest(BaseModel):
+    """Input payload for employee registration key issuance."""
 
-    refresh_token: str = Field(min_length=10, max_length=4096)
+    model_config = ConfigDict(extra="forbid")
 
+    target_role: StaffRoleClaim
+    ttl_seconds: int = Field(default=7 * 24 * 60 * 60, gt=0, le=30 * 24 * 60 * 60)
