+++ apps/backend/src/hrm_backend/admin/services/__init__.py
@@ -1,0 +1,5 @@
+"""Service layer for admin domain workflows."""
+
+from hrm_backend.admin.services.admin_service import AdminService
+
+__all__ = ["AdminService"]
