+++ apps/backend/src/hrm_backend/admin/utils/__init__.py
@@ -1,0 +1,1 @@
+"""Utility helpers for admin domain."""
