+++ apps/backend/src/hrm_backend/admin/utils/roles.py
@@ -1,0 +1,3 @@
+"""Role constants shared by admin domain services and schemas."""
+
+STAFF_ROLES = frozenset({"admin", "hr", "manager", "employee", "leader", "accountant"})
