+++ apps/backend/src/hrm_backend/auth/routers/__init__.py
@@ -1,6 +1,5 @@
-"""Routers for auth and admin API endpoints."""
+"""HTTP routers for auth domain APIs."""
 
-from hrm_backend.auth.routers.admin_v1 import router as admin_router
 from hrm_backend.auth.routers.v1 import router as auth_router
 
-__all__ = ["auth_router", "admin_router"]
+__all__ = ["auth_router"]
