+++ apps/backend/src/hrm_backend/auth/schemas/__init__.py
@@ -1,19 +1,12 @@
-"""Pydantic schemas for auth/admin requests, responses, and token claims."""
+"""Pydantic schemas for auth requests, responses, and token claims."""
 
 from hrm_backend.auth.schemas.requests import (
-    AdminCreateEmployeeKeyRequest,
-    AdminCreateStaffRequest,
-    AdminStaffUpdateRequest,
     LoginRequest,
     RefreshRequest,
     RegisterRequest,
 )
 from hrm_backend.auth.schemas.responses import (
-    AdminStaffListItem,
-    AdminStaffListResponse,
-    EmployeeRegistrationKeyResponse,
     MeResponse,
-    StaffResponse,
     TokenResponse,
 )
 from hrm_backend.auth.schemas.token_claims import AuthContext, TokenClaims
@@ -22,15 +15,8 @@ __all__ = [
     "RegisterRequest",
     "LoginRequest",
     "RefreshRequest",
-    "AdminCreateStaffRequest",
-    "AdminStaffUpdateRequest",
-    "AdminCreateEmployeeKeyRequest",
     "TokenResponse",
     "MeResponse",
-    "StaffResponse",
-    "AdminStaffListItem",
-    "AdminStaffListResponse",
-    "EmployeeRegistrationKeyResponse",
     "TokenClaims",
     "AuthContext",
 ]
