+++ docs/architecture/decisions.md
@@ -26,6 +26,7 @@ Use this log for decisions that change interfaces, data models, deployment topol
 | ADR-0019 | 2026-03-05 | accepted | Freeze OpenAPI contract and enforce drift checks in CI | architect + backend-engineer | API governance, CI/CD, frontend integration |
 | ADR-0020 | 2026-03-05 | accepted | Introduce ADMIN-01 frontend route guard and admin shell with Sentry tags | architect + frontend-engineer | frontend routing, access control UX, observability |
 | ADR-0021 | 2026-03-05 | accepted | Add ADMIN-02 staff management list/update APIs with strict admin safety guard | architect + backend-engineer + frontend-engineer | admin API contracts, RBAC, audit trail, frontend admin workspace |
+| ADR-0022 | 2026-03-05 | accepted | Extract admin governance flows into dedicated `admin` backend package | architect + backend-engineer | backend package boundaries, maintainability, test topology |
 
 ## ADR-0001
 - Context: Project is at bootstrap stage and lacks durable knowledge artifacts.
@@ -304,3 +305,16 @@ Use this log for decisions that change interfaces, data models, deployment topol
   - Admin safety invariants are enforced centrally in business logic.
   - Audit evidence is expanded with `admin.staff:list` and `admin.staff:update` success/failure traces.
   - `ADMIN-03` (employee key lifecycle UI/API expansion) remains isolated as next phase.
+
+## ADR-0022
+- Context: Admin governance flows were implemented in `hrm_backend/auth`, which broke bounded-context separation and conflicted with the package structure baseline used by other domains (for example `candidates`).
+- Decision:
+  - Extract admin HTTP/business/persistence layers to dedicated package `hrm_backend/admin`.
+  - Keep package structure aligned with extraction-ready baseline:
+    `models`, `schemas`, `services`, `dao`, `routers`, `utils`, `dependencies`, plus explicit `infra` adapters.
+  - Keep API contracts unchanged (`/api/v1/admin/*`, reason-code behavior, RBAC requirements), but route orchestration through `admin` service/DAO layers.
+  - Keep `hrm_backend/auth` scoped to auth/session lifecycle only (`register/login/refresh/logout/me`).
+- Consequences:
+  - Domain boundaries are explicit and consistent with engineering standards.
+  - Admin feature development no longer couples to auth service internals.
+  - Test topology now follows package flow (`tests/unit/admin`, `tests/integration/admin`).
